// deploy_ios/src/app/models/user.ts
export class User {
    id: string;
    login: string = "";
    password: string = "demo";
    rpassword: string = "";
    role: string = "user";
    application: string = "playclips";
    email: string = "";
    auth: string = "1004";
    username: string = "";
    isUserSelected: boolean = false;
}